+++ src/main/java/com/thehyundai/thepet/domain/order/OrderServiceImpl.java
@@ -4,8 +4,8 @@ import com.thehyundai.thepet.domain.cart.CartService;
 import com.thehyundai.thepet.domain.cart.CartVO;
 import com.thehyundai.thepet.global.exception.BusinessException;
 import com.thehyundai.thepet.global.exception.ErrorCode;
-import com.thehyundai.thepet.global.EntityValidator;
-import com.thehyundai.thepet.global.TableStatus;
+import com.thehyundai.thepet.global.util.EntityValidator;
+import com.thehyundai.thepet.global.cmcode.TableStatus;
 import com.thehyundai.thepet.domain.product.ProductService;
 import com.thehyundai.thepet.domain.product.ProductVO;
 import com.thehyundai.thepet.domain.subscription.CurationMapper;
