+++ src/main/java/com/thehyundai/thepet/global/sms/HcSmsEventListener.java
@@ -1,4 +1,4 @@
-package com.thehyundai.thepet.sms;
+package com.thehyundai.thepet.global.sms;
 
 import com.thehyundai.thepet.domain.heendycar.HcBranchVO;
 import com.thehyundai.thepet.domain.heendycar.HcReservationVO;
